requirements.txt
# Core data processing
pandas>=2.0.0

# Google Cloud dependencies
google-cloud-storage>=2.10.0
google-auth>=2.22.0
gcsfs>=2023.6.0

# Date and timezone handling
pytz>=2023.3

# Environment configuration
python-dotenv>=1.1.1

# Parquet support for pandas
pyarrow>=12.0.0